requirements.txt
crewai==0.30.11
crewai-tools==0.2.6
langchain-google-genai==1.0.3
python-dotenv==1.0.1
PyPDF2